add file download component
